Move RXCROPMATURITYSHARED into rxcropmaturity.py.
